PythonProjects/AI_agent/advanced_agent/pyproject.toml
[project]
name = "advanced-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firecrawl-py>=4.18.0",
    "langchain>=1.2.10",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
